Passes kernel config from VMM
