+++ src/core/Cargo.toml
@@ -10,7 +10,7 @@ crate-type = ["staticlib"]
 bitfield-struct = "0.8.0"
 humansize = "2.1.3"
 libc = "0.2.155"
-obconf = { path = "../obconf" }
+obconf = { path = "../obconf", features = ["serde"] }
 obfw = { git = "https://github.com/obhq/firmware-dumper.git", features = ["read", "std"] }
 obvirt = { path = "../obvirt" }
 param = { path = "../param" }
