+++ src/obconf/Cargo.toml
@@ -4,3 +4,4 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+serde = { version = "1.0.210", features = ["derive"], default-features = false, optional = true }
