+++ src/obconf/src/lib.rs
@@ -14,6 +14,16 @@ pub enum BootEnv {
 
 /// Runtime configurations for the kernel.
 #[repr(C)]
+#[derive(Clone)]
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 pub struct Config {
     pub max_cpu: NonZero<usize>,
+}
+
+impl Default for Config {
+    fn default() -> Self {
+        Self {
+            max_cpu: NonZero::new(1).unwrap(),
+        }
+    }
 }
